Refactor faq component
