+++ .eslintrc.js
@@ -23,6 +23,7 @@ module.exports = {
     'no-use-before-define': 'off',
     'import/prefer-default-export': 'off',
     'react/require-default-props': 'off',
+    'max-len': 'off',
   },
   settings: {
     'import/resolver': {
