+++ src/components/Tabs/styles.ts
